Update Hero.jsx - Changed wrapping of buttons in hero section
Wrapping of buttons in hero section was changed such that the button tag was wrapped inside link tag instead of link inside button as it was in the past. This eliminated the issue of the button doing nothing when user clicked the space of padding inside the button instead of navigating to the specified page.